api/requirements.txt
eventlet==0.33.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask_Login==0.6.1
Flask_SocketIO==5.1.2
Flask_SQLAlchemy==2.5.1
paho_mqtt==1.6.1
python-dotenv==0.20.0
schedule==1.1.0
Werkzeug==2.1.2
mariadb==1.0.11
gunicorn==20.1.0